huge dialog util refactor
