+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/app/models/playback/managers/HQDialogManager.java
@@ -1,23 +1,17 @@
 package com.liskovsoft.smartyoutubetv2.common.app.models.playback.managers;
 
-import android.content.Context;
-import android.os.Build;
-import com.liskovsoft.sharedutils.helpers.Helpers;
 import com.liskovsoft.sharedutils.helpers.MessageHelpers;
 import com.liskovsoft.smartyoutubetv2.common.R;
 import com.liskovsoft.smartyoutubetv2.common.app.models.playback.PlayerEventListenerHelper;
-import com.liskovsoft.smartyoutubetv2.common.app.models.playback.controller.PlaybackEngineController;
 import com.liskovsoft.smartyoutubetv2.common.app.models.playback.ui.OptionCategory;
 import com.liskovsoft.smartyoutubetv2.common.app.models.playback.ui.OptionItem;
 import com.liskovsoft.smartyoutubetv2.common.app.models.playback.ui.UiOptionItem;
 import com.liskovsoft.smartyoutubetv2.common.app.presenters.AppDialogPresenter;
 import com.liskovsoft.smartyoutubetv2.common.app.views.ViewManager;
 import com.liskovsoft.smartyoutubetv2.common.autoframerate.FormatItem;
-import com.liskovsoft.smartyoutubetv2.common.autoframerate.FormatItem.VideoPreset;
-import com.liskovsoft.smartyoutubetv2.common.misc.AppDataSourceManager;
 import com.liskovsoft.smartyoutubetv2.common.prefs.PlayerData;
+import com.liskovsoft.smartyoutubetv2.common.utils.AppDialogUtil;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -28,10 +22,6 @@ public class HQDialogManager extends PlayerEventListenerHelper {
     private static final String TAG = HQDialogManager.class.getSimpleName();
     private static final int VIDEO_FORMATS_ID = 132;
     private static final int AUDIO_FORMATS_ID = 133;
-    private static final int VIDEO_BUFFER_ID = 134;
-    private static final int BACKGROUND_PLAYBACK_ID = 135;
-    private static final int VIDEO_PRESETS_ID = 136;
-    private static final int AUDIO_DELAY_ID = 137;
     // NOTE: using map, because same item could be changed time to time
     private final Map<Integer, OptionCategory> mCategories = new LinkedHashMap<>();
     private final Map<Integer, OptionCategory> mCategoriesInt = new LinkedHashMap<>();
@@ -105,12 +95,12 @@ public class HQDialogManager extends PlayerEventListenerHelper {
     }
 
     private void addVideoBufferCategory() {
-        addCategoryInt(createVideoBufferCategory(getActivity(), mPlayerData,
+        addCategoryInt(AppDialogUtil.createVideoBufferCategory(getActivity(), mPlayerData,
                 () -> getController().restartEngine()));
     }
 
     private void addAudioDelayCategory() {
-        addCategoryInt(createAudioShiftCategory(getActivity(), mPlayerData,
+        addCategoryInt(AppDialogUtil.createAudioShiftCategory(getActivity(), mPlayerData,
                 () -> getController().restartEngine()));
     }
 
@@ -132,13 +122,13 @@ public class HQDialogManager extends PlayerEventListenerHelper {
 
     private void addBackgroundPlaybackCategory() {
         OptionCategory category =
-                createBackgroundPlaybackCategory(getActivity(), mPlayerData, this::updateBackgroundPlayback);
+                AppDialogUtil.createBackgroundPlaybackCategory(getActivity(), mPlayerData, this::updateBackgroundPlayback);
 
         addCategoryInt(category);
     }
 
     private void addPresetsCategory() {
-        addCategoryInt(createVideoPresetsCategory(
+        addCategoryInt(AppDialogUtil.createVideoPresetsCategory(
                 getActivity(), mPlayerData, () -> getController().setFormat(mPlayerData.getFormat(FormatItem.TYPE_VIDEO))));
     }
 
@@ -183,131 +173,5 @@ public class HQDialogManager extends PlayerEventListenerHelper {
                     break;
             }
         }
-    }
-
-    public static OptionCategory createBackgroundPlaybackCategory(Context context, PlayerData playerData) {
-        return createBackgroundPlaybackCategory(context, playerData, () -> {});
-    }
-
-    private static OptionCategory createBackgroundPlaybackCategory(Context context, PlayerData playerData, Runnable onSetCallback) {
-        String categoryTitle = context.getString(R.string.category_background_playback);
-
-        List<OptionItem> options = new ArrayList<>();
-        options.add(UiOptionItem.from(context.getString(R.string.option_background_playback_off),
-                optionItem -> {
-                    playerData.setBackgroundMode(PlaybackEngineController.BACKGROUND_MODE_DEFAULT);
-                    onSetCallback.run();
-                }, playerData.getBackgroundMode() == PlaybackEngineController.BACKGROUND_MODE_DEFAULT));
-        if (Helpers.isAndroidTV(context) && Build.VERSION.SDK_INT < 26) { // useful only for pre-Oreo UI
-            options.add(UiOptionItem.from(context.getString(R.string.option_background_playback_behind) + " (Android TV 5,6,7)",
-                    optionItem -> {
-                        playerData.setBackgroundMode(PlaybackEngineController.BACKGROUND_MODE_PLAY_BEHIND);
-                        onSetCallback.run();
-                    }, playerData.getBackgroundMode() == PlaybackEngineController.BACKGROUND_MODE_PLAY_BEHIND));
-        }
-        if (Helpers.isPictureInPictureSupported(context)) {
-            options.add(UiOptionItem.from(context.getString(R.string.option_background_playback_pip),
-                    optionItem -> {
-                        playerData.setBackgroundMode(PlaybackEngineController.BACKGROUND_MODE_PIP);
-                        onSetCallback.run();
-                    }, playerData.getBackgroundMode() == PlaybackEngineController.BACKGROUND_MODE_PIP));
-        }
-        options.add(UiOptionItem.from(context.getString(R.string.option_background_playback_only_audio),
-                optionItem -> {
-                    playerData.setBackgroundMode(PlaybackEngineController.BACKGROUND_MODE_SOUND);
-                    onSetCallback.run();
-                }, playerData.getBackgroundMode() == PlaybackEngineController.BACKGROUND_MODE_SOUND));
-
-        return OptionCategory.from(BACKGROUND_PLAYBACK_ID, OptionCategory.TYPE_RADIO, categoryTitle, options);
-    }
-
-    public static OptionCategory createVideoPresetsCategory(Context context, PlayerData playerData) {
-        return createVideoPresetsCategory(context, playerData, () -> {});
-    }
-
-    private static OptionCategory createVideoPresetsCategory(Context context, PlayerData playerData, Runnable onFormatSelected) {
-        return OptionCategory.from(
-                VIDEO_PRESETS_ID,
-                OptionCategory.TYPE_RADIO,
-                context.getString(R.string.title_video_presets),
-                fromPresets(
-                        context,
-                        AppDataSourceManager.instance().getVideoPresets(),
-                        playerData,
-                        onFormatSelected
-                )
-        );
-    }
-
-    private static List<OptionItem> fromPresets(Context context, VideoPreset[] presets, PlayerData playerData, Runnable onFormatSelected) {
-        List<OptionItem> result = new ArrayList<>();
-
-        FormatItem selectedFormat = playerData.getFormat(FormatItem.TYPE_VIDEO);
-        boolean isPresetSelection = selectedFormat != null && selectedFormat.isPreset();
-
-        for (VideoPreset preset : presets) {
-            result.add(0, UiOptionItem.from(preset.name,
-                    option -> setFormat(preset.format, playerData, onFormatSelected),
-                    isPresetSelection && preset.format.equals(selectedFormat)));
-        }
-
-        result.add(0, UiOptionItem.from(
-                context.getString(R.string.video_preset_disabled),
-                optionItem -> setFormat(playerData.getDefaultVideoFormat(), playerData, onFormatSelected),
-                !isPresetSelection));
-
-        return result;
-    }
-
-    public static OptionCategory createVideoBufferCategory(Context context, PlayerData playerData) {
-        return createVideoBufferCategory(context, playerData, () -> {});
-    }
-
-    private static void setFormat(FormatItem formatItem, PlayerData playerData, Runnable onFormatSelected) {
-        if (playerData.isLegacyCodecsForced()) {
-            playerData.forceLegacyCodecs(false);
-        }
-        playerData.setFormat(formatItem);
-        onFormatSelected.run();
-    }
-
-    private static OptionCategory createVideoBufferCategory(Context context, PlayerData playerData, Runnable onBufferSelected) {
-        String videoBufferTitle = context.getString(R.string.video_buffer);
-        List<OptionItem> optionItems = new ArrayList<>();
-        optionItems.add(createVideoBufferOption(context, playerData, R.string.video_buffer_size_low, PlaybackEngineController.BUFFER_LOW, onBufferSelected));
-        optionItems.add(createVideoBufferOption(context, playerData, R.string.video_buffer_size_med, PlaybackEngineController.BUFFER_MED, onBufferSelected));
-        optionItems.add(createVideoBufferOption(context, playerData, R.string.video_buffer_size_high, PlaybackEngineController.BUFFER_HIGH, onBufferSelected));
-        return OptionCategory.from(VIDEO_BUFFER_ID, OptionCategory.TYPE_RADIO, videoBufferTitle, optionItems);
-    }
-
-    private static OptionItem createVideoBufferOption(Context context, PlayerData playerData, int titleResId, int type, Runnable onBufferSelected) {
-        return UiOptionItem.from(
-                context.getString(titleResId),
-                optionItem -> {
-                    playerData.setVideoBufferType(type);
-                    onBufferSelected.run();
-                },
-                playerData.getVideoBufferType() == type);
-    }
-
-    public static OptionCategory createAudioShiftCategory(Context context, PlayerData playerData) {
-        return createAudioShiftCategory(context, playerData, () -> {});
-    }
-
-    private static OptionCategory createAudioShiftCategory(Context context, PlayerData playerData, Runnable onSetCallback) {
-        String title = context.getString(R.string.audio_shift);
-
-        List<OptionItem> options = new ArrayList<>();
-
-        for (int delayMs : Helpers.range(-4_000, 4_000, 50)) {
-            options.add(UiOptionItem.from(context.getString(R.string.audio_shift_sec, Helpers.toString(delayMs / 1_000f)),
-                    optionItem -> {
-                        playerData.setAudioDelayMs(delayMs);
-                        onSetCallback.run();
-                    },
-                    delayMs == playerData.getAudioDelayMs()));
-        }
-
-        return OptionCategory.from(AUDIO_DELAY_ID, OptionCategory.TYPE_RADIO, title, options);
     }
 }
