+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/app/models/playback/managers/PlayerUIManager.java
@@ -1,6 +1,5 @@
 package com.liskovsoft.smartyoutubetv2.common.app.models.playback.managers;
 
-import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
 import android.view.KeyEvent;
@@ -8,10 +7,10 @@ import com.liskovsoft.mediaserviceinterfaces.MediaItemManager;
 import com.liskovsoft.mediaserviceinterfaces.MediaService;
 import com.liskovsoft.mediaserviceinterfaces.data.MediaItem;
 import com.liskovsoft.mediaserviceinterfaces.data.MediaItemMetadata;
-import com.liskovsoft.sharedutils.helpers.Helpers;
 import com.liskovsoft.sharedutils.helpers.KeyHelpers;
 import com.liskovsoft.sharedutils.helpers.MessageHelpers;
 import com.liskovsoft.sharedutils.mylogger.Log;
+import com.liskovsoft.sharedutils.rx.RxUtils;
 import com.liskovsoft.smartyoutubetv2.common.R;
 import com.liskovsoft.smartyoutubetv2.common.app.models.data.Video;
 import com.liskovsoft.smartyoutubetv2.common.app.models.playback.PlayerEventListenerHelper;
@@ -25,26 +24,19 @@ import com.liskovsoft.smartyoutubetv2.common.app.presenters.ChannelPresenter;
 import com.liskovsoft.smartyoutubetv2.common.app.presenters.SearchPresenter;
 import com.liskovsoft.smartyoutubetv2.common.app.presenters.dialogs.menu.VideoMenuPresenter;
 import com.liskovsoft.smartyoutubetv2.common.autoframerate.FormatItem;
-import com.liskovsoft.smartyoutubetv2.common.exoplayer.other.SubtitleManager.OnSelectSubtitleStyle;
-import com.liskovsoft.smartyoutubetv2.common.exoplayer.other.SubtitleManager.SubtitleStyle;
 import com.liskovsoft.smartyoutubetv2.common.misc.MotherActivity;
 import com.liskovsoft.smartyoutubetv2.common.prefs.PlayerData;
-import com.liskovsoft.sharedutils.rx.RxUtils;
 import com.liskovsoft.smartyoutubetv2.common.prefs.SearchData;
-import com.liskovsoft.smartyoutubetv2.common.utils.Utils;
+import com.liskovsoft.smartyoutubetv2.common.utils.AppDialogUtil;
 import com.liskovsoft.youtubeapi.service.YouTubeMediaService;
 import io.reactivex.Observable;
 
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 
 public class PlayerUIManager extends PlayerEventListenerHelper implements MetadataListener {
     private static final String TAG = PlayerUIManager.class.getSimpleName();
     private static final long SUGGESTIONS_RESET_TIMEOUT_MS = 500;
-    private static final int SUBTITLE_STYLES_ID = 45;
     private final Handler mHandler;
     private final MediaItemManager mMediaItemManager;
     private final VideoLoaderManager mVideoLoader;
@@ -321,7 +313,7 @@ public class PlayerUIManager extends PlayerEventListenerHelper implements Metada
 
         dialogPresenter.clear();
 
-        Utils.appendShareDialogItems(getActivity(), dialogPresenter, getController().getVideo());
+        AppDialogUtil.appendShareDialogItems(getActivity(), dialogPresenter, getController().getVideo());
 
         dialogPresenter.showDialog(getController().getVideo().title);
     }
@@ -336,10 +328,10 @@ public class PlayerUIManager extends PlayerEventListenerHelper implements Metada
 
     @Override
     public void onVideoZoomClicked() {
-        OptionCategory videoZoomCategory = createVideoZoomCategory(
+        OptionCategory videoZoomCategory = AppDialogUtil.createVideoZoomCategory(
                 getActivity(), mPlayerData, () -> getController().setVideoZoomMode(mPlayerData.getVideoZoomMode()));
 
-        OptionCategory videoAspectCategory = createVideoAspectCategory(
+        OptionCategory videoAspectCategory = AppDialogUtil.createVideoAspectCategory(
                 getActivity(), mPlayerData, () -> getController().setVideoAspectRatio(mPlayerData.getVideoAspectRatio()));
 
         AppDialogPresenter settingsPresenter = AppDialogPresenter.instance(getActivity());
@@ -506,88 +498,5 @@ public class PlayerUIManager extends PlayerEventListenerHelper implements Metada
 
     private interface MediaItemObservable {
         Observable<Void> call(MediaItem item);
-    }
-
-    public static OptionCategory createSubtitleStylesCategory(Context context, PlayerData playerData) {
-        return createSubtitleStylesCategory(context, playerData, style -> {});
-    }
-
-    private static OptionCategory createSubtitleStylesCategory(Context context, PlayerData playerData, OnSelectSubtitleStyle onSelectSubtitleStyle) {
-        List<SubtitleStyle> subtitleStyles = playerData.getSubtitleStyles();
-        
-        String subtitleStyleTitle = context.getString(R.string.subtitle_style);
-
-        return OptionCategory.from(SUBTITLE_STYLES_ID, OptionCategory.TYPE_RADIO, subtitleStyleTitle, fromSubtitleStyles(context, playerData, subtitleStyles, onSelectSubtitleStyle));
-    }
-
-    private static List<OptionItem> fromSubtitleStyles(Context context, PlayerData playerData, List<SubtitleStyle> subtitleStyles, OnSelectSubtitleStyle onSelectSubtitleStyle) {
-        List<OptionItem> styleOptions = new ArrayList<>();
-
-        for (SubtitleStyle subtitleStyle : subtitleStyles) {
-            styleOptions.add(UiOptionItem.from(
-                    context.getString(subtitleStyle.nameResId),
-                    option -> {
-                        playerData.setSubtitleStyle(subtitleStyle);
-                        onSelectSubtitleStyle.onSelectSubtitleStyle(subtitleStyle);
-                    },
-                    subtitleStyle.equals(playerData.getSubtitleStyle())));
-        }
-
-        return styleOptions;
-    }
-
-    public static OptionCategory createVideoZoomCategory(Context context, PlayerData playerData) {
-        return createVideoZoomCategory(context, playerData, () -> {});
-    }
-
-    private static OptionCategory createVideoZoomCategory(Context context, PlayerData playerData, Runnable onSelectZoomMode) {
-        List<OptionItem> options = new ArrayList<>();
-
-        for (int[] pair : new int[][] {
-                {R.string.video_zoom_default, PlaybackEngineController.ZOOM_MODE_DEFAULT},
-                {R.string.video_zoom_fit_width, PlaybackEngineController.ZOOM_MODE_FIT_WIDTH},
-                {R.string.video_zoom_fit_height, PlaybackEngineController.ZOOM_MODE_FIT_HEIGHT},
-                {R.string.video_zoom_fit_both, PlaybackEngineController.ZOOM_MODE_FIT_BOTH},
-                {R.string.video_zoom_stretch, PlaybackEngineController.ZOOM_MODE_STRETCH}}) {
-            options.add(UiOptionItem.from(context.getString(pair[0]),
-                    optionItem -> {
-                        playerData.setVideoZoomMode(pair[1]);
-                        onSelectZoomMode.run();
-                    },
-                    playerData.getVideoZoomMode() == pair[1]));
-        }
-
-        String videoZoomTitle = context.getString(R.string.video_zoom);
-
-        return OptionCategory.from(SUBTITLE_STYLES_ID, OptionCategory.TYPE_RADIO, videoZoomTitle, options);
-    }
-
-    private static OptionCategory createVideoAspectCategory(Context context, PlayerData playerData, Runnable onSelectAspectMode) {
-        List<OptionItem> options = new ArrayList<>();
-
-        Map<String, Float> pairs = new LinkedHashMap<>();
-        pairs.put(context.getString(R.string.video_zoom_default), PlaybackEngineController.ASPECT_RATIO_DEFAULT);
-        pairs.put("1:1", PlaybackEngineController.ASPECT_RATIO_1_1);
-        pairs.put("4:3", PlaybackEngineController.ASPECT_RATIO_4_3);
-        pairs.put("5:4", PlaybackEngineController.ASPECT_RATIO_5_4);
-        pairs.put("16:9", PlaybackEngineController.ASPECT_RATIO_16_9);
-        pairs.put("16:10", PlaybackEngineController.ASPECT_RATIO_16_10);
-        pairs.put("21:9 (2.33:1)", PlaybackEngineController.ASPECT_RATIO_21_9);
-        pairs.put("64:27 (2.37:1)", PlaybackEngineController.ASPECT_RATIO_64_27);
-        pairs.put("2.21:1", PlaybackEngineController.ASPECT_RATIO_221_1);
-        pairs.put("2.35:1", PlaybackEngineController.ASPECT_RATIO_235_1);
-        pairs.put("2.39:1", PlaybackEngineController.ASPECT_RATIO_239_1);
-
-        for (Entry<String, Float> entry: pairs.entrySet()) {
-            options.add(UiOptionItem.from(entry.getKey(),
-                    optionItem -> {
-                        playerData.setVideoAspectRatio(entry.getValue());
-                        onSelectAspectMode.run();
-                    }, Helpers.floatEquals(playerData.getVideoAspectRatio(), entry.getValue())));
-        }
-
-        String videoZoomTitle = context.getString(R.string.video_aspect);
-
-        return OptionCategory.from(SUBTITLE_STYLES_ID, OptionCategory.TYPE_RADIO, videoZoomTitle, options);
     }
 }
